pptx_indexer/requirements.txt
# Core dependencies
python-pptx>=0.6.21
networkx>=3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM & Embeddings
openai>=1.0.0
anthropic>=0.7.0
sentence-transformers>=2.2.0

# Vector stores
chromadb>=0.3.0

# OCR & Image Processing
pytesseract>=0.3.10
Pillow>=9.0.0

# Utilities
python-dotenv>=0.21.0
numpy>=1.24.0
loguru>=0.7.0

# Optional: LLM providers
mistralai>=0.0.7
replicate>=0.12.0

# Optional: Vector stores
pinecone-client>=2.2.0
weaviate-client>=3.20.0

# Optional: Graph databases
neo4j>=5.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
